requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pyshp
pyproj
typer
gunicorn
asyncpg
psycopg2-binary
flask-cors
Flask
PyJWT
pandas
SQLAlchemy>=2.0
Flask-Migrate
openpyxl
xlsxwriter
cloudinary
python-dotenv





